RELATED: BB-2139 codemirror component
