+++ package.json
@@ -119,7 +119,7 @@
   "dependencies": {
     "@formatjs/intl-pluralrules": "^1.3.6",
     "@gooddata/gooddata-js": "^13.0.1",
-    "@gooddata/goodstrap": "^68.9.0",
+    "@gooddata/goodstrap": "^68.10.0",
     "@gooddata/js-utils": "^3.10.5",
     "@gooddata/numberjs": "^3.2.4",
     "@gooddata/typings": "^2.24.2",
