+++ styles/scss/main.scss
@@ -8,7 +8,8 @@ $gdc-goodstrap-basepath: "~@gooddata/goodstrap/lib/";
 @import "./bulletChart.scss";
 @import "./attributeFilter.scss";
 @import "./measureValueFilter.scss";
-@import "./measureNumberFormat/styles.scss";
+@import "./measureNumberFormat.scss";
+@import "./syntaxHighlightingInput.scss";
 
 /* hide IE10+ eraser */
