+++ styles/scss/measureNumberFormat.scss
@@ -1,6 +1,5 @@
 // (C) 2020 GoodData Corporation
 @import "~@gooddata/goodstrap/lib/core/styles/variables.scss";
-@import "./syntaxHighlightingInput.scss";
 
 .gd-measure-number-format-dropdown-body {
     padding: 5px 0;
