+++ styles/scss/syntaxHighlightingInput.scss
@@ -45,6 +45,10 @@ $cm-color-keyword: #ab55a3;
         }
     }
 
+    .CodeMirror-empty {
+        color: $gd-color-disabled;
+    }
+
     .CodeMirror-code {
         .CodeMirror-matchingbracket {
             font-weight: bold;
